+++ CHANGELOG.md
@@ -14,11 +14,19 @@ Features:
 
 * Enable users to configure session-expiration on browser-close via env-variable or config-file [#119](https://github.com/O-X-L/ansible-webui/issues/119)
 * Save/handle user-language on the backend [#116](https://github.com/O-X-L/ansible-webui/issues/116)
+* Email-plugin handles username=email [#122](https://github.com/O-X-L/ansible-webui/issues/122)
 
 
 Fixes:
 
 * Improved page-loading [#95](https://github.com/O-X-L/ansible-webui/issues/95)
+* Ensure users do not get alerted multiple times for a single execution [#80](https://github.com/O-X-L/ansible-webui/issues/80)
+* Do not trigger git-repo-update for all jobs in job-list-view - only when job is opened [#123](https://github.com/O-X-L/ansible-webui/issues/123)
+
+
+Chore:
+
+* Bumped Django version from 5.1 to 5.2
 
 ----